pyproject.toml
[tool.poetry]
name = "scgdi-motor50cv"
version = "0.1.0"
description = "Servidor OPC UA + MQTT para monitoramento de motor trifásico (exercício final)"
authors = ["Seu Nome <voce@exemplo.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
asyncua = "^1.1.6"
gmqtt = "^0.6.13"
pydantic = "^2.8.2"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
uvloop = {version = "^0.20.0", platform = "linux"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
